Fix for bug #9725 -- inconsistent quality metrics.  The problem here was that the verdict quality metrics needed to include the surface normal when computing some metrics (scaled jacobian).  This was already being done for triangle metrics, so I followed the same approach for quads.
A ripe bug I found while making these changes was in QualityController::check_shape_of_quads.  There, a call to QuadQuality::metric_value(quad,QuadQuality::SHEAR) was being made.  I fixed it so that it calls SHAPE instead.  This caused a few tests to fail, because now when meshing, the user is apprised of through an error, the first element that fails the shape metric.  

This change also caused some problems when meshing with the mirror/copy/morph scheme.  With these schemes, Cubit copies a mesh and the elements can be initially reversed on the surface.   Previously this was ok because these reversed elements were smoothed, agnostic to the fact that they were reversed.  Now that the quality knows about surface normals, the smoothing fails where previously it did not.  The fix was to, if necessary, reverse normals before smoothing.


git-svn-id: file:///scratch/cjstimp/svn/cubit/CUBIT_SOURCE/trunk@389954 d353b8dc-a10e-11dd-9ca7-d543032e456e